Change migration order and increase int

+++ database/migrations/2017_05_01_170000_create_machine_jobs_table.php
@@ -16,7 +16,7 @@ class CreateMachineJobsTable extends Migration
         Schema::create('machine_jobs', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('machine_id')->unsigned();
-            $table->unsignedTinyInteger('duration');
+            $table->unsignedSmallInteger('duration');
             $table->boolean('completed')->default(false);
             $table->timestamps();
